rotate button for second form
