+++ Freelance-Helper.code-workspace
@@ -10,6 +10,11 @@
 	"settings": {
 		"files.autoSave": "onFocusChange",
 		"vs-code-prettier-eslint.prettierLast": false,
-		"git.ignoreLimitWarning": true
+		"git.ignoreLimitWarning": true,
+		"jest.disabledWorkspaceFolders": [
+			"frontend",
+		
+		],
+		
 	}
 }
