+++ frontend/.eslintrc.json
@@ -17,6 +17,7 @@
     "es2020": true
   },
   "rules": {
-    "react/react-in-jsx-scope": "off"
+    "react/react-in-jsx-scope": "off",
+    "react/prop-types": "off"
   }
 }
