+++ frontend/src/components/SideBar.jsx
@@ -52,7 +52,7 @@ const SideBar = () => {
         <Box w="full">
           <Box>
             <Link
-              href="/freelanceHelper"
+              href="/"
               as={Link}
               display="flex"
               alignItems="center"
