+++ frontend/vite.config.js
@@ -4,7 +4,7 @@ import eslintPlugin from "vite-plugin-eslint";
 
 export default defineConfig({
   // depending on your application, base can also be "/"
-  base: "/dashboard",
+  base: "/",
   plugins: [
     react(),
     eslintPlugin({
